Refactor size constants
