--- src/constants/index.js
@@ -1,5 +1,0 @@
-export const SIZES = {
-  REGULAR: 'regular',
-  LARGE: 'large',
-  X_LARGE: 'x-large',
-}
