+++ src/constants/size.js
@@ -1,0 +1,4 @@
+export const REGULAR = 'regular'
+export const LARGE = 'large'
+export const X_LARGE = 'x-large'
+export const SIZES = [REGULAR, LARGE, X_LARGE]
